SoftwareVersion: Add git hash suffix with git-version crate
Since we add it to the bugreport!() macro, it will be evaluated in the
context of the dependant project, and give relevant git hash info.

Features:
* Optional via Cargo feature "git_hash"
* Prints '<no git>' as git hash if git is not on PATH
* Triggers rebuild if git hash (or index) change [1]

Note that since the bugreport!() macro is evaluated in the context of
the dependent project, it is not possible to use #[cfg()] inside of it.
So we have to have two different definitions of it depnding on feature
"git_hash".

[1] Via elegant include_bytes!() hack by git-version crate, see
https://github.com/fusion-engineering/rust-git-version/blob/v0.3.4/git-version-macro/src/lib.rs#L40