Silently ignore missing git repo even with git_hash feature
If an official release of a dependent project is built with e.g. `cargo
install <name>`, there will be no git repo available. But that is OK,
because an official release is built, so the git hash is uniquely
determined by the release git tag.

So to keep output clean and to-the-point, silently skip printing git
hash in this case.